Better UX

+++ out/commands/removeProject.js
@@ -1,4 +1,13 @@
 "use strict";
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
 var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
@@ -20,40 +29,47 @@ function removeProject(projectName, application = "default") {
             const pathApplication = settings.path;
             _dir = path_1.join(projectFolder, pathApplication, projectName);
         }
-        console.log(_dir);
         const removeTree = (dir) => {
             try {
                 fs_1.rmdirSync(dir, { recursive: true });
+                return true;
             }
             catch (error) {
                 console.error(error);
+                return false;
             }
         };
-        try {
-            inquirer_1.default
-                .prompt([
-                {
-                    type: "confirm",
-                    default: false,
-                    message: `Are you sure you want to delete locally the project ${projectName.bgGreen.white}`
-                        .red,
-                    name: "confirmDeletion",
-                },
-            ])
-                .then((answer) => {
-                if (answer.confirmDeletion) {
-                    removeTree(_dir);
-                    console.log(`${_dir} has been sucessfully removed!`.red);
-                    if (main_1.env.token)
-                        removeGithubProject(projectName);
-                }
-                else {
-                    console.log("Nothing has been deleted".green);
-                }
-            });
+        if (fs_1.existsSync(_dir)) {
+            console.log(_dir);
+            try {
+                inquirer_1.default
+                    .prompt([
+                    {
+                        type: "confirm",
+                        default: false,
+                        message: `Are you sure you want to delete locally the project ${projectName.bgGreen.white}`
+                            .red,
+                        name: "confirmDeletion",
+                    },
+                ])
+                    .then((answer) => {
+                    if (answer.confirmDeletion) {
+                        removeTree(_dir);
+                        console.log(`${_dir} has been sucessfully removed!`.red);
+                        if (main_1.env.token)
+                            removeGithubProject(projectName);
+                    }
+                    else {
+                        console.log("Nothing has been deleted".green);
+                    }
+                });
+            }
+            catch (error) {
+                console.log(error);
+            }
         }
-        catch (error) {
-            console.log(error);
+        else {
+            console.log(`The project ${projectName} at ${_dir} does not exist!`.red);
         }
     }
     catch (error) {
@@ -62,23 +78,28 @@ function removeProject(projectName, application = "default") {
 }
 exports.default = removeProject;
 function removeGithubProject(repo) {
-    inquirer_1.default
-        .prompt([
-        {
-            type: "confirm",
-            default: false,
-            message: "[GITHUB]\n".red +
-                "are you sure you want to delete this repository: " +
-                repo.green +
-                " ?",
-            name: "removeRepo",
-        },
-    ])
-        .then((data) => {
-        if (data.removeRepo)
-            github_1.deleteRepo(repo);
-        else
-            console.log("[GITHUB] ".red + "repository not deleted.".green);
+    return __awaiter(this, void 0, void 0, function* () {
+        const repoExists = yield github_1.repoAlreadyExists(repo);
+        if (repoExists) {
+            inquirer_1.default
+                .prompt([
+                {
+                    type: "confirm",
+                    default: false,
+                    message: "[GITHUB]\n".red +
+                        "are you sure you want to delete this repository: " +
+                        repo.green +
+                        " ?",
+                    name: "removeRepo",
+                },
+            ])
+                .then((data) => {
+                if (data.removeRepo)
+                    github_1.deleteRepo(repo);
+                else
+                    console.log("[GITHUB] ".red + "repository not deleted.".green);
+            });
+        }
     });
 }
 exports.removeGithubProject = removeGithubProject;
